add: post, put, delete comments errors

+++ src/comments/comments.controller.ts
@@ -8,37 +8,39 @@ export class CommentsController {
   constructor(private readonly commentsService: CommentsService) { }
 
   @Post('/api/v1/comments')
-  
-  create(@Body() createCommentDto: CreateCommentDto) {
-    const createComment = this.commentsService.create(createCommentDto);
+  async create(@Body() createCommentDto: CreateCommentDto) {
+    const createComment = await this.commentsService.create(createCommentDto);
+    if(createComment !== createCommentDto) throw new Error('400: invalid input, comment not created or updated')
     return createComment;
   }
 
   @Get('/api/v1/posts/:id/comments')
-  findAll() {
+  async findAll() {
     const findAllCommentsOfPost = this.commentsService.findAll();
     return findAllCommentsOfPost;
   }
   @Get('/api/v1/comments')
-  findAllComments() {
+  async findAllComments() {
     const findAllComments = this.commentsService.findAll();
     return findAllComments;
   }
 
   @Get('/api/v1/posts/:id/comments/:comment_id')
-  findOne(@Param('comment_id') comment_id: string) {
+  async findOne(@Param('comment_id') comment_id: string) {
     const findOneComment = this.commentsService.findOne(comment_id);
     return findOneComment;
   }
 
   @Put('/api/v1/posts/:id/comments/:comment_id')
-  update(@Param('comment_id') comment_id: string, @Body() updateCommentDto: UpdateCommentDto) {
+  async update(@Param('comment_id') comment_id: string, @Body() updateCommentDto: UpdateCommentDto) {
+    if(!this.commentsService.findOne(comment_id)) throw new Error('404: invalid input, comment not found')
     const updateComment = this.commentsService.update(comment_id, updateCommentDto);
     return updateComment;
   }
 
   @Delete('/api/v1/posts/:id/comments/:comment_id')
-  remove(@Param('comment_id') comment_id: string) {
+  async remove(@Param('comment_id') comment_id: string) {
+    if(!this.commentsService.findOne(comment_id)) throw new Error('404: invalid input, comment not found')
     const deleteComment = this.commentsService.remove(comment_id);
     return deleteComment;
   }
